+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/storage/ya.make
@@ -13,7 +13,6 @@ SRCS(
 
 GO_TEST_SRCS(
     storage_ydb_test.go
-    hanging_ydb_test.go
 )
 
 END()
